requirements.txt
pyfiglet
rich
discord
asyncpg
typing
pyyaml
humanize
git+https://github.com/Rapptz/discord-ext-menus